# docs/swagger.yaml
definitions:
  model.EditHouse:
    properties:
      address:
        description: 房屋地址
        type: string
      hr_type:
        description: 房屋类型 租还是售
        type: string
      id:
        type: string
      message:
        description: 房屋信息
        type: string
      phone:
        description: 联系方式
        type: string
      price:
        description: 房屋价格
        type: string
      publisher:
        description: 发布者
        type: string
    type: object
  model.HouseKeeping:
    properties:
      address:
        description: 地址
        type: string
      hk_type:
        description: 家政服务类型
        type: string
      id:
        type: integer
      phone:
        description: 联系方式
        type: string
      status:
        description: 服务状态
        type: string
      username:
        description: 用户账户
        type: string
    type: object
  model.HouseRentSale:
    properties:
      address:
        description: 房屋地址
        type: string
      hr_type:
        description: 房屋类型 租还是售
        type: string
      id:
        type: integer
      message:
        description: 房屋信息
        type: string
      phone:
        description: 联系方式
        type: string
      price:
        description: 房屋价格
        type: string
      publisher:
        description: 发布者
        type: string
    type: object
  model.Notice:
    properties:
      content:
        description: 公告内容
        type: string
      id:
        type: integer
      publish_time:
        description: 公告发布时间
        type: string
      publisher:
        description: 公告发布者
        type: string
      title:
        description: 公告发布标题
        type: string
    type: object
  model.TSBX:
    properties:
      address:
        description: 投诉报修者地址
        type: string
      cr_type:
        description: 投诉 报修
        type: string
      id:
        type: integer
      message:
        description: 投诉报修详细信息
        type: string
      phone:
        description: 投诉报修者电话
        type: string
      status:
        description: 未处理 处理中 已处理
        type: string
      username:
        description: 投诉报修者账户
        type: string
    type: object
  model.User:
    properties:
      address:
        description: 用户地址
        type: string
      house_id:
        description: 房屋ID
        type: string
      id:
        type: integer
      nickname:
        description: 用户昵称
        type: string
      password:
        description: 用户密码
        type: string
      phone:
        description: 用户电话
        type: string
      username:
        description: 用户账号
        type: string
    type: object
  model.UserRegister:
    properties:
      house_id:
        description: 房屋ID
        type: string
      nickname:
        description: 用户昵称
        type: string
      password:
        description: 用户密码
        type: string
      username:
        description: 用户账号
        type: string
    type: object
  util.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      message:
        type: string
    type: object
info:
  contact: {}
  license:
    name: zwy
  title: 社区便民服务接口
  version: "1.0"
paths:
  /admin/housekeeping:
    get:
      description: 获取全部的家政服务请求
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: getAllHouseKeeping
      tags:
      - 管理员模块
  /admin/login:
    get:
      description: 管理员登录
      parameters:
      - description: 管理员账号
        in: query
        name: username
        required: true
        type: string
      - description: 管理员密码
        in: query
        name: password
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: adminLogin
      tags:
      - 管理员模块
  /admin/notice:
    get:
      description: 获取全部社区公告信息
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: getAllNotice
      tags:
      - 管理员模块
  /admin/tsbx:
    get:
      description: 后台管理员获取全部的投诉报修
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: gtAllComplainRepair
      tags:
      - 管理员模块
  /housekeeping/add:
    post:
      description: 添加一个家政服务请求
      parameters:
      - description: 家政服务结构体
        in: body
        name: houseKeeping
        required: true
        schema:
          $ref: '#/definitions/model.HouseKeeping'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: addHouseKeeping
      tags:
      - 家政服务模块
  /housekeeping/get:
    get:
      description: 获取每个人申请的家政服务
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: getHouseKeepingByUsername
      tags:
      - 家政服务模块
  /houserentsale/all:
    get:
      description: 获取全部的房屋信息
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: RentSale
      tags:
      - 房屋租售模块
  /houserentsale/delete:
    get:
      description: 根据ID删除房屋
      parameters:
      - description: 房屋主键ID
        in: query
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: DeleteHouseRentSale
      tags:
      - 房屋租售模块
  /houserentsale/edit:
    post:
      description: 修改房屋信息
      parameters:
      - description: 修改房屋信息结构体
        in: body
        name: editHouse
        required: true
        schema:
          $ref: '#/definitions/model.EditHouse'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: editHouse
      tags:
      - 房屋租售模块
  /houserentsale/publish:
    post:
      description: 发布房屋租售信息
      parameters:
      - description: 房屋租售结构体
        in: body
        name: houseRent
        required: true
        schema:
          $ref: '#/definitions/model.HouseRentSale'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: publishHouse
      tags:
      - 房屋租售模块
  /houserentsale/rent:
    get:
      description: 获取全部租房信息
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: Rent
      tags:
      - 房屋租售模块
  /houserentsale/sale:
    get:
      description: 获取全部的售房信息
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: Sale
      tags:
      - 房屋租售模块
  /notice/delete:
    get:
      description: 删除公告根据标题
      parameters:
      - description: 公告标题
        in: query
        name: title
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: deleteNotice
      tags:
      - 社区公告模块
  /notice/detail:
    get:
      description: 获取公告的详细信息
      parameters:
      - description: 公告标题
        in: query
        name: title
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: detailNotice
      tags:
      - 社区公告模块
  /notice/publish:
    post:
      description: 发布社区公告
      parameters:
      - description: Notice结构体
        in: body
        name: notice
        required: true
        schema:
          $ref: '#/definitions/model.Notice'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: publishNotice
      tags:
      - 社区公告模块
  /notice/recommend:
    get:
      description: 获取推荐公告，按发布时间最新排序
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: recommendNotice
      tags:
      - 社区公告模块
  /property/get:
    get:
      description: 获取物业费
      parameters:
      - description: 房屋ID
        in: query
        name: houseId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: getProperty
      tags:
      - 物业费模块
  /property/pay:
    get:
      description: 缴纳物业费
      parameters:
      - description: 房屋ID
        in: query
        name: houseId
        required: true
        type: string
      - description: 缴纳金额
        in: query
        name: property
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: payProperty
      tags:
      - 物业费模块
  /tsbx/add:
    post:
      description: 添加投诉报修
      parameters:
      - description: TSBX结构体
        in: body
        name: tb
        required: true
        schema:
          $ref: '#/definitions/model.TSBX'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: addComplainRepair
      tags:
      - 投诉报修模块
  /tsbx/get:
    get:
      description: 获取本人的投诉报修
      parameters:
      - description: 用户账号
        in: query
        name: username
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: getComplainRepairByUsername
      tags:
      - 投诉报修模块
  /user/get:
    get:
      description: 获取用户信息根据username
      parameters:
      - description: 用户账户
        in: query
        name: username
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: getUserInfo
      tags:
      - 用户模块
  /user/info:
    post:
      description: 修改用户信息
      parameters:
      - description: User结构体包含要修改的字段
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: editUserInfo
      tags:
      - 用户模块
  /user/login:
    get:
      description: 用户使用账户和密码登录
      parameters:
      - description: 用户账户
        in: query
        name: username
        required: true
        type: string
      - description: 用户密码
        in: query
        name: password
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: userLogin
      tags:
      - 用户模块
  /user/password:
    post:
      description: 修改用户密码
      parameters:
      - description: User结构体只包含username和password
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: editUserPassword
      tags:
      - 用户模块
  /user/register:
    get:
      description: 用户输入在账号，密码，并绑定自己的房屋ID，昵称
      parameters:
      - description: 用户注册结构体
        in: body
        name: userRegister
        required: true
        schema:
          $ref: '#/definitions/model.UserRegister'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.Response'
      summary: userRegister
      tags:
      - 用户模块
swagger: "2.0"