# router/user.go
package router

import (
	"github.com/gin-gonic/gin"
	"graduationproject/model"
	"graduationproject/util"
	"net/http"
)

//@Summary userLogin
//@Description 用户使用账户和密码登录
//@Tags 用户模块
//@Param username query string true "用户账户"
//@Param password query string true "用户密码"
//@Success 200 {object} util.Response
//@Failure 500 {object} util.Response
//@Router /user/login [get]
func userLogin(context *gin.Context) {
	var username = context.Query("username")
	var password = context.Query("password")
	err := model.Login(username, password)
	if err != nil {
		context.JSON(http.StatusOK, util.Response{
			Code:    555,
			Message: util.RequestFail,
			Data:    nil,
		})
	}
	context.JSON(http.StatusOK, util.Response{
		Code:    666,
		Message: util.LoginSuccess,
		Data:    nil,
	})
}

//@Summary userRegister
//@Description 用户输入在账号，密码，并绑定自己的房屋ID，昵称
//@Tags 用户模块
//@Param userRegister body model.UserRegister true "用户注册结构体"
//@Success 200 {object} util.Response
//@Failure 500 {object} util.Response
//@Router /user/register [get]
func userRegister(c *gin.Context) {
	var ur model.UserRegister
	if err := c.ShouldBindJSON(&ur); err != nil {
		c.JSON(200, util.Response{
			Code:    555,
			Message: util.RequestFail,
			Data:    nil,
		})
		return
	}
	err := model.Register(ur)
	if err != nil {
		c.JSON(200, util.Response{
			Code:    555,
			Message: util.RequestFail,
			Data:    nil,
		})
		return
	}
	c.JSON(http.StatusOK, util.Response{
		Code:    666,
		Message: util.RegisterSuccess,
		Data:    nil,
	})

}

//@Summary editUserInfo
//@Description 修改用户信息
//@Tags 用户模块
//@Param user body model.User true "User结构体包含要修改的字段"
//@Success 200 {object} util.Response
//@Failure 500 {object} util.Response
//@Router /user/info [post]
func editUserInfo(c *gin.Context) {
	var u model.User
	if err := c.ShouldBindJSON(&u); err != nil {
		c.JSON(200, util.Response{
			Code:    555,
			Message: util.RequestFail,
			Data:    nil,
		})
		return
	}
	err := model.EditInfo(u)
	if err != nil {
		c.JSON(200, util.Response{
			Code:    555,
			Message: util.RequestFail,
			Data:    nil,
		})
		return
	}
	c.JSON(200, util.Response{
		Code:    555,
		Message: util.InfoSuccess,
		Data:    nil,
	})

}

//@Summary editUserPassword
//@Description 修改用户密码
//@Tags 用户模块
//@Param user body model.User true "User结构体只包含username和password"
//@Success 200 {object} util.Response
//@Failure 500 {object} util.Response
//@Router /user/password [post]
func editUserPassword(c *gin.Context) {
	var u model.User
	if err := c.ShouldBindJSON(&u); err != nil {
		c.JSON(200, util.Response{
			Code:    555,
			Message: util.RequestFail,
			Data:    nil,
		})
	}
	err := model.EditPassword(u)
	if err != nil {
		c.JSON(200, util.Response{
			Code:    555,
			Message: util.RequestFail,
			Data:    nil,
		})
	}
	c.JSON(200, util.Response{
		Code:    666,
		Message: util.PasswordSuccess,
		Data:    nil,
	})

}

//@Summary getUserInfo
//@Description 获取用户信息根据username
//@Tags 用户模块
//@Param username query string true "用户账户"
//@Success 200 {object} util.Response
//@Failure 500 {object} util.Response
//@Router /user/get [get]
func getInfo(c *gin.Context) {
	query := c.Query("username")
	info, err := model.GetInfo(query)
	if err != nil {
		c.JSON(http.StatusOK, util.Response{
			Code:    555,
			Message: util.RequestFail,
			Data:    info,
		})
		return
	}
	c.JSON(http.StatusOK, util.Response{
		Code:    666,
		Message: util.GetDataSuccess,
		Data:    info,
	})
}

func UserRouters(engine *gin.Engine) {
	group := engine.Group("/user")
	{
		group.GET("/get", getInfo)
		group.GET("/login", userLogin)
		group.POST("/register", userRegister)
		group.POST("/info", editUserInfo)
		group.POST("/password", editUserPassword)
	}
}